プレイヤーのマップ移動処理の追加2
- マップ移動したら、描画される背景が移動したマップの物に
切り替わる処理の追加

- ワールドマップを定義する場所の追加
  - js/game/common_function/world_map.js // ワールドマップ全体
  - js/game/common_function/world_map_parts //ワールドマップの各パーツ

branch#05
