+++ readme.md
@@ -7,7 +7,10 @@
             * arrow.js <!-- プレイヤーの武器である弓矢 Arrow クラス -->
             * player.js <!-- プレイヤー Player クラス -->
         - common_function <!-- ゲーム内容における処理全般で使える関数 -->
+            - world_map_parts <!-- マップのパーツ(16 × 16 サイズ = 1画面分) -->
             * change_screen_to.js <!-- メインループを止めて、次の画面に遷移する関数 -->
+            * draw_map.js <!-- マップのパーツのデータに従って、キャンパスに描画する関数 -->
+            * world_map.js <!-- マップのパーツを組み合わせて、世界全体(ワールドマップ)を定義 -->
         * start.js <!-- スタート画面 -->
         * setting.js <!-- 設定画面 -->
         * game.js <!-- プレイ画面 -->
@@ -19,3 +22,4 @@
     * main.js <!-- 一番最初に呼ばれる関数 -->
     * call_main.js <!-- main 関数を呼ぶための関数(html と main 関数の仲介役) -->
 * index.html <!-- 初期ページ -->
+* readme.md <!-- このファイル -->
